+++ daemon/src/engine/renderer/tr_scene.cpp
@@ -617,7 +617,7 @@ void RE_RenderScene( const refdef_t *fd )
 	R_AddWorldLightsToScene();
 
 	// derived info
-	tr.refdef.floatTime = tr.refdef.time * 0.001f;
+	tr.refdef.floatTime = float(double(tr.refdef.time) * 0.001);
 
 	tr.refdef.numDrawSurfs = r_firstSceneDrawSurf;
 	tr.refdef.drawSurfs = backEndData[ tr.smpFrame ]->drawSurfs;
